build(bazel): fix broken aio github links on Windows
Caused the aio e2e tests to fail.
